Moved commands to subfolders sorted by help category
