+++ commands/fun/neko.js
@@ -1,6 +1,6 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { MessageEmbed } = require('discord.js');
-const { objToString, randomColor } = require('../functions');
+const { objToString, randomColor } = require('../../functions');
 const client = require('nekos.life');
 const neko = new client();
