+++ commands/fun/shards.js
