+++ commands/moderator/ban.js
@@ -1,6 +1,6 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { Permissions } = require('discord.js');
-const { errembed, permcheck, punembed } = require('../functions');
+const { errembed, permcheck, punembed } = require('../../functions');
 
 module.exports = {
 	data: new SlashCommandBuilder()
