+++ commands/utility/avatar.js
@@ -1,6 +1,6 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { MessageEmbed } = require('discord.js');
-const { randomColor } = require('../functions');
+const { randomColor } = require('../../functions');
 
 module.exports = {
 	data: new SlashCommandBuilder()
