+++ commands/utility/ping.js
